requirements.txt
numpy
pandas
matplotlib
kaggle
python-dotenv